refactor: extract remotecontrol plugin API policy service
